thermal: Remove the core control switch
This shouldn't even have been there not being CONFIG_MSM_CORE_CTL there, so it doesn't make any sense to wrap it under an ifdef while the main part isn't there.
I still haven't understood why it was there and was shown but anyway it isn't needed and it isn't supposed to be there.

Signed-off-by: dd3boh <dade.garberi@gmail.com>